leaderboard: split sync classes and remove wrapper entrypoints
